requirements.txt
Flask==3.1.2
mysql-connector-python==9.4.0
openpyxl==3.1.5
pandas==2.3.2
